+++ app/src/androidTest/java/com/example/gnatovskiy_lesson_2/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.gnatovskiy_lesson_1
+package com.example.gnatovskiy_lesson_2
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.gnatovskiy_lesson_1", appContext.packageName)
+        assertEquals("com.example.gnatovskiy_lesson_2", appContext.packageName)
     }
 }
