+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.gnatovskiyer_lesson_1">
+    package="com.example.gnatovskiy_lesson_2">
 
     <application
         android:allowBackup="true"
@@ -8,16 +8,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.GnatovskiyER_lesson_1">
+        android:theme="@style/Theme.Gnatovskiy_lesson_2" >
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".FirstTaskActivity"></activity>
-        <activity android:name=".SecondTaskActivity"></activity>
-
+        <activity android:name=".FirstActivity"/>
+        <activity android:name=".SecondActivity"/>
     </application>
-
 </manifest>
