--- app/src/main/java/com/example/gnatovskiy_lesson_1/data/Students.kt
@@ -1,11 +1,0 @@
-package com.example.gnatovskiyer_lesson_1.data
-
-import java.time.Year
-
-data class Students (
-    val id : Int,
-    val name : String,
-    val surname : String,
-    val grade : String,
-    val birthdayYear: Year
-)
