+++ app/src/main/java/com/example/gnatovskiy_lesson_2/SecondActivity.kt
@@ -1,0 +1,11 @@
+package com.example.gnatovskiy_lesson_2
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class SecondActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.second_activity)
+    }
+}
