+++ app/src/main/res/drawable-land/ic_pushkin_signature.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+android:width="194dp"
+android:height="57dp"
+android:viewportWidth="194"
+android:viewportHeight="57">
+<path
+    android:pathData="M4.473,23.796c-2.208,0.539 -6.65,3.459 -3.21,5.606c3.504,2.188 11.528,-3.989 14.056,-5.819c5.099,-3.691 9.887,-7.848 14.708,-11.889c2.479,-2.077 4.946,-4.168 7.415,-6.257c1.394,-1.18 3.541,-3.847 5.326,-4.413c-0.364,-0.072 -0.729,-0.144 -1.093,-0.216c0.411,1.333 -3.155,5.1 -3.845,6.03c-1.394,1.88 -2.867,3.703 -4.353,5.511c-3.042,3.703 -6.212,7.311 -9.563,10.736c-1.389,1.419 -2.843,2.817 -4.458,3.978c-1.835,1.319 0.248,-2.594 0.619,-3.134c1.051,-1.531 2.393,-2.883 3.812,-4.072c-1.038,0.869 -2.376,3.089 -3.125,4.741c-3.163,6.978 6.265,3.337 9.02,1.925c5.699,-2.92 13.901,-7.673 17.095,-13.485c0.439,-0.799 -1.497,-0.578 -1.797,-0.033c-2.692,4.9 -9.12,8.557 -13.699,11.373c-1.187,0.73 -10.651,6.02 -9.12,0.927c0.5,-1.664 3.575,-4.589 3.103,-6.473c-0.042,-0.165 -0.231,-0.227 -0.373,-0.26c-2.776,-0.651 -5.952,4.217 -7.19,6.021c-1.485,2.163 -1.319,5.252 1.885,3.589c4.823,-2.504 8.617,-7.689 12.125,-11.708c2.316,-2.654 4.545,-5.38 6.714,-8.156c1.079,-1.38 5.712,-5.993 5.084,-8.029c-0.12,-0.388 -0.816,-0.304 -1.093,-0.217c-1.988,0.63 -3.715,2.604 -5.269,3.917c-3.521,2.975 -7.037,5.957 -10.574,8.914c-3.616,3.024 -7.25,6.027 -10.973,8.92c-1.199,0.932 -2.415,1.848 -3.694,2.667c-1.89,1.211 -3.879,2.25 -5.918,3.186c-6.253,2.872 -4.008,-2.284 -1.357,-2.931C5.896,24.461 5.632,23.513 4.473,23.796L4.473,23.796z"
+    android:fillColor="#231F20"/>
+<path
+    android:pathData="M52.941,25.763c-0.093,-1.522 2.155,-3.644 3.083,-4.607c3.924,-4.07 9.026,-7.211 13.606,-10.482c1.588,-1.134 3.209,-2.255 4.963,-3.121c3.767,-1.86 0.383,3.763 0.071,4.23c-2.952,4.418 -6.937,8.274 -10.963,11.706c-0.854,0.728 1.078,0.709 1.549,0.316c2.606,-2.174 5.334,-4.208 8.179,-6.058c1.102,-0.717 6.818,-3.658 4.482,-0.494c-1.14,1.544 -2.705,2.759 -3.942,4.223c-0.612,0.725 -2.997,3.76 -0.617,4.14c2.625,0.419 5.222,-1.458 7.304,-2.79c1.649,-1.055 3.229,-2.223 4.825,-3.358c0.615,-0.438 1.237,-0.865 1.868,-1.282c1.366,-0.904 -1.119,-0.108 -1.526,0.139c-2.138,1.296 -5.354,3.419 -3.608,6.188c1.163,1.843 4.942,-0.066 6.151,-0.715c2.237,-1.201 2.551,-2.477 3.303,-4.442c-0.644,0.127 -1.287,0.253 -1.932,0.38c0.119,2.063 -3.365,5.558 -4.51,6.988c-4.474,5.591 -9.784,10.608 -14.993,15.504c-3.862,3.63 -7.851,7.145 -12.039,10.394c-0.083,0.064 -4.354,3.238 -4.542,2.769c-0.329,-0.826 1.294,-2.457 1.736,-2.99c8.245,-9.941 21.531,-16.422 32.345,-23.103c4.268,-2.637 8.567,-5.225 12.769,-7.968c1.562,-1.02 4.164,-2.338 5.044,-4.111c0.151,-0.303 -0.005,-0.461 -0.308,-0.544c-2.461,-0.677 -9.351,7.895 -7.158,8.939c2.618,1.247 15.291,-5.626 14.318,-7.689c-0.281,-0.595 -1.385,-0.212 -1.741,0.101c-1.221,1.073 -4.709,7.146 -1.65,7.771c3.988,0.815 9.769,-4.447 12.793,-6.355c2.238,-1.413 -0.084,0.18 -0.976,1.561c-0.425,0.658 -2.665,5.215 -0.233,3.698c2.481,-1.548 5.073,-2.934 7.658,-4.298c1.411,-0.746 2.844,-1.449 4.313,-2.071c1.507,-0.638 -0.536,1.366 -1.092,2.133c-1.429,1.973 -3.872,3.818 -4.935,5.992c-0.313,0.641 0.668,0.583 1.028,0.503c3.435,-0.756 7.338,-4.985 10.24,-6.884c1.014,-0.663 2.052,-1.281 3.108,-1.872c1.989,-1.111 2.182,-1.532 -0.029,-0.626c-1.849,0.758 -4.166,1.861 -5.361,3.538c-1.508,2.114 -2.086,5.19 1.193,5.463c4.945,0.412 10.486,-5.357 13.657,-8.272c-0.558,-0.06 -1.116,-0.119 -1.674,-0.178c-0.718,0.824 -5.516,6.448 -2.656,7.15c1.761,0.432 3.269,-0.229 4.79,-1.087c1.364,-0.769 2.521,-2.021 3.665,-3.063c0.043,-0.856 -0.192,-0.949 -0.705,-0.279c-0.508,0.397 -0.675,0.716 -0.977,1.261c-1.959,3.542 5.197,1.291 5.773,1.012c2.386,-1.154 4.596,-2.683 6.84,-4.082c1.395,-0.87 3.393,-2.201 1.735,0.362c-0.813,1.258 -2.096,2.785 -3.416,3.585c0.522,0.055 1.044,0.11 1.565,0.165c0.209,-0.56 12.5,-6.799 4.935,-1.29c-2.498,1.819 -1.087,4.403 1.508,4.936c1.816,0.372 3.866,-1.796 5.031,-2.817c2.205,-1.934 7.171,-8.644 10.31,-8.535c0.729,0.025 2.06,-1.049 0.752,-1.095c-3.454,-0.12 -6.654,3.731 -8.9,5.842c-1.385,1.302 -6.199,7.345 -8.025,4.181c-1.146,-1.985 4.745,-3.895 3.641,-6.109c-0.157,-0.314 -0.53,-0.293 -0.834,-0.275c-2.142,0.124 -9.61,3.43 -10.348,5.405c-0.26,0.697 1.364,0.287 1.565,0.165c1.122,-0.681 7.632,-7.676 3.547,-7.162c-2.08,0.262 -4.174,2.009 -5.871,3.078c-1.814,1.142 -3.625,2.288 -5.558,3.222c-1.36,0.657 -2.299,-1.081 -1.17,-1.966c0.348,-0.272 0.489,-0.375 0.73,-0.703c0.132,-0.179 0.074,-0.399 -0.131,-0.487c-1.657,-0.716 -3.231,0.983 -4.323,1.977c-0.163,0.148 -5.214,4.114 -4.988,1.812c0.091,-0.935 0.821,-1.819 1.316,-2.583c0.588,-0.908 1.287,-1.762 1.997,-2.578c0.576,-0.662 -1.326,-0.498 -1.674,-0.178c-1.212,1.114 -13.576,11.395 -12.544,5.229c0.746,-4.465 5.648,-3.676 8.036,-6.084c0.181,-0.182 0.322,-0.493 -0.005,-0.625c-2.183,-0.884 -5.207,1.438 -6.881,2.469c-2.776,1.711 -7.677,6.478 -10.125,7.017c0.343,0.167 0.686,0.335 1.028,0.503c0.429,-0.876 9.402,-10.382 5.035,-9.461c-2.828,0.596 -5.555,2.377 -8.072,3.729c-2.272,1.22 -5.255,2.409 -7.146,4.215c-0.289,0.276 -0.229,0.6 0.183,0.683c2.907,0.587 5.159,-5.716 6.205,-7.415c0.064,-0.105 0.133,-0.303 -0.002,-0.398c-1.822,-1.284 -4.861,1.604 -6.25,2.493c-2.461,1.576 -5.022,3.412 -7.818,4.332c-1.201,0.396 0.92,-5.613 1.737,-6.332c-0.58,0.033 -1.16,0.067 -1.74,0.101c0.338,0.715 -5.684,4.098 -6.42,4.519c-0.605,0.347 -5.563,3.037 -4.221,0.377c0.643,-1.273 1.774,-2.411 2.73,-3.437c0.631,-0.677 1.306,-1.313 1.991,-1.933c1.145,-1.036 -1.387,-0.035 -1.88,0.392c-2.359,2.043 -5.229,3.615 -7.864,5.263c-8.775,5.484 -17.638,10.821 -26.265,16.539c-5.594,3.707 -11.613,7.815 -15.746,13.196c-0.836,1.088 -1.937,3.652 0.552,3.218c4.102,-0.716 7.958,-4.812 10.974,-7.355c6.799,-5.734 13.344,-11.951 19.305,-18.554c2.09,-2.315 8.26,-7.711 8.044,-11.449c-0.041,-0.723 -1.759,-0.07 -1.932,0.38c-0.043,0.114 -0.033,0.269 -0.026,0.387c0.043,0.398 -0.022,0.789 -0.195,1.171c-0.319,0.354 -0.64,0.708 -0.959,1.063c-1.423,1.672 -2.954,1.403 -4.595,-0.807c1.027,-2.009 3.732,-2.374 5.124,-3.994c0.11,-0.129 0.083,-0.273 -0.003,-0.398c-1.207,-1.773 -7.165,3.556 -7.944,4.1c-1.087,0.759 -2.2,1.476 -3.333,2.165c-1.739,1.057 -4.435,1.351 -2.543,-1.169c1.547,-2.061 3.79,-3.561 5.102,-5.806c1.859,-3.184 -4.662,-1.065 -5.395,-0.71c-4.07,1.975 -7.807,5.002 -11.269,7.89c0.516,0.105 1.033,0.211 1.549,0.316c4.538,-3.868 10.383,-8.977 12.777,-14.605c1.243,-2.922 -1.04,-3.737 -3.56,-2.702c-6.487,2.664 -12.803,8.168 -18.116,12.609c-1.736,1.451 -5.507,4.489 -5.342,7.173C51.048,26.907 52.987,26.505 52.941,25.763L52.941,25.763z"
+    android:fillColor="#231F20"/>
+<path
+    android:pathData="M180.259,20.53c1.289,7.046 -3.061,11.833 -5.984,17.671c-0.11,0.221 1.665,0.11 1.9,-0.36c2.962,-5.915 7.314,-10.729 6.016,-17.829C182.127,19.663 180.181,20.102 180.259,20.53L180.259,20.53z"
+    android:fillColor="#231F20"/>
+<path
+    android:pathData="M166.69,43.271c5.008,-3.083 10.359,-5.203 15.614,-7.769c2.954,-1.443 5.435,-3.342 8.457,-4.722c1.209,-0.552 2.567,-1.062 3.198,-2.306c0.025,-0.09 0.005,-0.168 -0.063,-0.234c-0.049,-0.071 -0.098,-0.141 -0.146,-0.211c-0.344,-0.493 -2.217,0.135 -1.801,0.733c-0.18,-0.258 -14.74,7.895 -14.767,7.905c-4.069,1.579 -7.964,3.821 -11.673,6.104C164.4,43.456 166.111,43.627 166.69,43.271L166.69,43.271z"
+    android:fillColor="#231F20"/>
+</vector>
