+++ app/src/main/res/drawable/button_normal.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners android:topLeftRadius="7dp"
+        android:bottomRightRadius="7dp"/>
+    <size android:height="20dp"
+        android:width="100dp"/>
+    <solid android:color="#d46610"/>
+</shape>
