+++ app/src/main/res/layout/main_activity.xml
@@ -2,17 +2,17 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
+    <Button
+        android:id="@+id/button_first"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/button_first_task"
+        android:layout_margin="20dp"/>
+    <Button
+        android:id="@+id/button_second"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/button_first_task"
+        android:layout_margin="20dp"/>
 
-    <Button
-        android:id="@+id/buttonFirstTask"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        android:text="FirstTask"/>
-    <Button
-        android:id="@+id/buttonSecondTask"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="20dp"
-        android:text="SecondTask"/>
 </LinearLayout>
