+++ app/src/main/res/layout/second_activity.xml
@@ -1,55 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-        <EditText
-            android:id="@+id/edit_text_student"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:imeOptions="actionSend"
-            android:inputType="text"/>
-        <Button
-            android:id="@+id/button_View_student"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="20dp"
-            android:text="View Students"
-            />
-    </LinearLayout>
+    android:layout_height="match_parent"
+    android:gravity="center_horizontal"
+    android:background="@drawable/fon">
 
-    <LinearLayout
-        android:layout_width="match_parent"
+    <TextView
+        android:id="@+id/image_quadrs"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="horizontal">
-        <TextView
-            android:id="@+id/view_id_student"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginHorizontal="10dp"/>
-        <TextView
-            android:id="@+id/view_name_student"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginHorizontal="10dp"/>
-        <TextView
-            android:id="@+id/view_surname_student"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginHorizontal="10dp"/>
-        <TextView
-            android:id="@+id/view_grade_student"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginHorizontal="10dp"/>
-        <TextView
-            android:id="@+id/view_birthday_year"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginHorizontal="10dp"/>
-    </LinearLayout>
+        android:background="@drawable/two"
+        />
+    <androidx.appcompat.widget.AppCompatButton
+        android:id="@+id/button_click"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginVertical="20dp"
+        android:background="@drawable/button"/>
 </LinearLayout>
