+++ app/src/main/res/values-v21/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="text_version">Радость и счастье </string>
+    <!--    <string name=""-->
+</resources>
