+++ app/src/main/res/values-w325dp/colors.xml
@@ -6,9 +6,6 @@
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
-    <color name="red">#FF0000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="text_color">#FF0000</color>
-    <color name="button_color_1">#D46610</color>
-    <color name="green">#00ff00 </color>
+    <color name="text_color">#FF000000</color>
 </resources>
