+++ app/src/main/res/values/values-land/themes.xml
