+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "gnatovskiy_lesson_1"
+rootProject.name = "gnatovskiy_lesson_2"
